requirements_txt.txt
discord.py==2.3.2
requests==2.31.0
flask==2.3.3